refactor(filtereddropdown): extract styled components
