+++ src/components/molecules/DropdownFiltered/DropdownFiltered.tsx
@@ -1,28 +1,16 @@
 import Downshift, { ControllerStateAndHelpers, DownshiftProps } from 'downshift';
 import React from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { colors } from '../../../constants/colors';
-import { typography } from '../../../constants/typography';
 import ErrorMessage from '../../atoms/ErrorMessage/ErrorMessage';
 import InputLabel from '../../atoms/InputLabel/InputLabel';
-import InputText from '../../atoms/InputText/InputText';
 import SizedContainer from '../../layout/SizedContainer/SizedContainer';
 import Icon from '../../atoms/Icon/Icon';
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons';
 import { IField, IInput } from '../../types';
-
-export interface ISearchInputProps {
-  isOpen?: boolean;
-}
-
-export interface IOptionsListProps {
-  optionsListMaxHeight?: string;
-}
-
-export interface IOption extends React.HTMLAttributes<HTMLDivElement> {
-  highLighted?: boolean;
-  selected?: boolean;
-}
+import Option from './Option';
+import Options, { IOptionsListProps } from './Options';
+import { ISearchInputProps, SearchInput, SearchArrowWrap, SearchInputWrap } from './SearchInput';
 
 export interface IDropdownItem {
   value: string;
@@ -43,71 +31,6 @@ export interface IDropdownFilteredProps
    */
   items: IDropdownItem[];
 }
-
-const SearchInput = styled(InputText)<ISearchInputProps & IInput>`
-  margin: 0;
-  padding: 8px 32px 8px 16px;
-  ${({ hasError }) => !hasError && 'margin-bottom: 0'};
-  ${({ isOpen }) =>
-    isOpen &&
-    css`
-      box-shadow: 0 4px 1px 2px rgba(28, 33, 57, 0.15);
-      border-radius: 6px 6px 0 0;
-      border-color: ${colors.actionPlain};
-      border-bottom: 0;
-
-      /* Hack to simulate a border in the bottom in the input as :after
-         pseudo-elements doesn't work with inputs */
-      background: linear-gradient(${colors.greyLightest}, ${colors.greyLightest});
-      background-size: 95% 1px;
-      background-position: bottom center;
-      background-repeat: no-repeat;
-      &:focus {
-        border-bottom: 0;
-      }
-    `};
-`;
-
-const Option = styled.div<IOption>`
-  cursor: pointer;
-  padding: 8px;
-  line-height: 32px;
-  font-size: ${typography.sizes.text.body};
-  font-weight: 600;
-  color: ${colors.greyDarkest};
-  ${({ selected, highLighted }) =>
-    (selected || highLighted) &&
-    css`
-      color: ${colors.white};
-      background-color: ${colors.actionPlain};
-    `};
-`;
-
-const Options = styled.div<IOptionsListProps>`
-  z-index: 1;
-  width: 100%;
-  background: ${colors.white};
-  border: 2px solid ${colors.actionPlain};
-  border-top: 0;
-  position: absolute;
-  overflow: auto;
-  border-radius: 0 0 6px 6px;
-  max-height: ${({ optionsListMaxHeight = 'auto' }) => optionsListMaxHeight};
-  box-shadow: 0 2px 1px 2px rgba(28, 33, 57, 0.15);
-`;
-
-const SearchInputWrap = styled.div`
-  position: relative;
-`;
-
-const SearchArrowWrap = styled.div`
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  right: 12px;
-  cursor: pointer;
-  display: flex;
-`;
 
 const FieldError = styled(ErrorMessage)`
   margin-top: 5px;
